3 commit

+++ app/src/main/java/ru/korolevss/currencyapp/model/CurrencyItem.kt
@@ -6,6 +6,6 @@ data class CurrencyItem(
     val code: String,
     val name: String,
     val flagResource: Int,
-    val nominal: BigDecimal,
-    val value: BigDecimal
+    val nominal: Double,
+    val value: Double
 )
